Extract saved game into a separate component
