+++ src/App.js
@@ -1,14 +1,13 @@
 import { useState, useEffect, useCallback } from 'react';
 import './App.css';
 import Cell from './components/Cell'
+import SavedGameDropdown from './components/SavedGameDropdown';
 import { resetGrid, randomize, calculateNextState, ROW_SIZE, COL_SIZE } from './utils/helpers';
 
 const App = () => {
   const [simulate, setSimulate] = useState(false);
   const [triggerCount, setTriggerCount] = useState(0);
   const [allCellData, setAllCellData] = useState(resetGrid())
-  const [savedGames, setSavedGames] = useState();
-  const [savedGameToLoad, setSavedGameToLoad] = useState();
 
   const renderGrid = useCallback(() => {
     console.log("Rendering Grid")
@@ -20,16 +19,6 @@ const App = () => {
   }, [allCellData, simulate])
 
   useEffect(() => {
-    if(localStorage.getItem("savedGames") === null) {
-      localStorage.setItem("savedGames", JSON.stringify([]))
-      setSavedGames([])
-    } else {
-      setSavedGames(JSON.parse(localStorage.getItem("savedGames")))
-    }
-  }, [])
-
-
-  useEffect(() => {
     const interval = setInterval(() => {
       if (simulate) {
         calculateNextState(allCellData, setAllCellData)
@@ -39,16 +28,9 @@ const App = () => {
     return () => clearInterval(interval);
   }, [allCellData, simulate]);
 
-  const saveSavedGame= () => {
-    const key = `savedGame${localStorage.getItem("savedGames").length + 1}`
-    localStorage.setItem("savedGames", JSON.stringify([...JSON.parse(localStorage.getItem("savedGames")), key]))
-    localStorage.setItem(key, JSON.stringify(allCellData));
+  const clearBoard = () => {
+    setAllCellData(resetGrid())
   }
-
-  const loadSavedGame = (v) => {
-    setAllCellData(JSON.parse(localStorage.getItem(v)));
-  }
-
 
   return (
     <div className="App">
@@ -59,17 +41,8 @@ const App = () => {
       </table>
       <button onClick={() => setSimulate(!simulate)}>{simulate ? "Stop" : "Simulate!"}</button>
       <button onClick={() => setAllCellData(randomize(allCellData))}>Randomize!</button>
-      <button onClick={() => setAllCellData(resetGrid())}>Clear!</button>
-      <button onClick={() => saveSavedGame()}>Save as preset</button>
-      <button onClick={() => loadSavedGame()}>Load preset</button>
-      {
-        savedGames && <select name="savedGames" onChange={(e) => loadSavedGame(e.target.value)}>
-          <option selected disabled hidden value=''></option>
-          {
-            savedGames.map((savedGame) => <option value={savedGame}>{savedGame}</option>)
-          }
-        </select>
-      }
+      <button onClick={() => clearBoard()}>Clear!</button>
+      <SavedGameDropdown setAllCellData={setAllCellData} resetGrid={resetGrid} allCellData={allCellData}/>
 
       <div>
         Have been triggered {triggerCount} times
